backend/requirements.txt
uvicorn
google-cloud-aiplatform
websockets
google-genai
fastapi